Move NodeRemovalClusterStateTaskExecutor out of Zen Discovery
